requirements.txt
discord.py
feedparser
apscheduler
python-dotenv
beautifulsoup4